src/python_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.4.2
python-dotenv==1.0.0
loguru==0.7.2
uuid==1.30
aiohttp==3.9.1
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1 